Separate table data and db schema services
