+++ src/app/common/components/dataset-form/dataset-form.component.ts
@@ -5,6 +5,7 @@ import { MatButtonModule } from "@angular/material/button";
 import { MatInputModule } from "@angular/material/input";
 
 import { DataService } from "~services/data/data.service";
+import { SchemaService } from "~services/table/schema.service";
 import { DatasetFormInterface } from "./dataset-form.interface";
 
 /**
@@ -30,7 +31,7 @@ export class DatasetFormComponent implements OnInit {
   protected allOptions: string[] = [];
   protected filteredOptions: string[] = [];
 
-  constructor(private dataService: DataService) {}
+  constructor(private schemaService: SchemaService, private dataService: DataService) {}
 
   /** Loads valid table names and set up form controls. */
   ngOnInit(): void {
@@ -47,7 +48,7 @@ export class DatasetFormComponent implements OnInit {
 
   /** Fetches valid table names from the API. */
   private initializeOptions(): void {
-    this.dataService.getTableNames().subscribe(
+    this.schemaService.getTableNames().subscribe(
       names => {
         this.allOptions = names;
         this.filteredOptions = [...this.allOptions];
