+++ src/app/common/services/data/data.service.ts
@@ -1,61 +1,51 @@
 import { HttpParams, HttpResponse } from "@angular/common/http";
 import { Injectable } from "@angular/core";
-import { map, Observable, shareReplay } from "rxjs";
+import { map, Observable } from "rxjs";
 
 import { APIService } from "~services/api/api.service";
-import { PaginatedData, Schema } from "./data.interface";
+import { PaginatedData } from "./data.interface";
 
 /** Service for fetching data from the application database via the API. */
 @Injectable({
   providedIn: "root"
 })
 export class DataService {
-  private apiSchemaEndpoint: string = "meta/schema/";
-  private dbSchema$!: Observable<Schema>;
+  tableData$!: Observable<PaginatedData>;
 
   /**
    * Instantiate an observable for the database schema fetched from the API.
    * The API response is automatically cached for the lifetime of the service.
    */
-  constructor(private apiService: APIService) {
-    this.dbSchema$ = this.apiService.get(this.apiSchemaEndpoint).pipe(
-      map((response: HttpResponse<Object>) => response.body as Schema),
-      shareReplay(1)
-    );
-  }
-
-  /**
-   * Returns a list of table names from the database.
-   * @returns An observable containing the database schema.
-   */
-  getTableNames(): Observable<string[]> {
-    return this.dbSchema$.pipe(
-      map((schema: Schema) => Object.keys(schema.tables))
-    );
-  }
-
-  /**
-   * Returns the column names for a given database table.
-   * @param tableName The name of the table to fetch columns for.
-   * @returns An observable containing table column names.
-   */
-  getColumnNames(tableName: string): Observable<string[]> {
-    return this.dbSchema$.pipe(
-      map((schema: Schema) => Object.keys(schema["tables"][tableName]["columns"]))
-    );
-  }
+  constructor(private apiService: APIService) {}
 
   /**
    * Returns data from a database table by name.
    * @param tableName The name of the table to fetch data from.
-   * @param options Optional pagination/ordering settings for the returned data.
+   * @param options Pagination/ordering settings for the returned data.
    * @returns An observable containing table data.
    */
-  getTableData(
+  fetchTableData(
     tableName: string,
     options: { pageIndex?: number; pageSize?: number; orderBy?: string; direction?: string } = {}
-  ): Observable<PaginatedData> {
+  ): void {
 
+    const url: string = `db/${tableName}/`;
+    const params: HttpParams = this.buildRequestParams(options);
+
+    this.tableData$ = this.apiService.get(url, params).pipe(
+      map((response: HttpResponse<Object>) => ({
+        pageData: response.body,
+        tableLength: Number(response.headers.get("x-pagination-total"))
+      }))
+    );
+  }
+
+  /**
+   * Generates HTTP query parameters for paginated data requests.
+   * @param options Pagination/ordering settings.
+   * @returns HTTP query parameters suitable for paginating data.
+   */
+  private buildRequestParams(options: { pageIndex?: number; pageSize?: number; orderBy?: string; direction?: string }) {
     const {pageIndex = 0, pageSize = 0, orderBy, direction} = options;
     let params: HttpParams = new HttpParams({
       fromObject: {
@@ -66,12 +56,6 @@ export class DataService {
 
     if (orderBy) params = params.set("_order_by_", orderBy);
     if (direction) params = params.set("_direction_", direction);
-
-    return this.apiService.get(`db/${tableName}/`, params).pipe(
-      map((response: HttpResponse<Object>) => ({
-        pageData: response.body,
-        tableLength: Number(response.headers.get("x-pagination-total"))
-      }))
-    );
+    return params;
   }
 }
